refactor :: 테이블 셀 표시 로직 분리
- 헤더, 값에 따라 셀을 다르게 표시하는 로직을 `utils/table` 폴더로 분리
