+++ src/pages/management/InfoPage.tsx
@@ -5,7 +5,7 @@ import CustomInput from "../../components/CustomInput";
 import MainHeader from "../../components/MainHeader";
 import { isValidEmail, isValidPhone } from "../../utils/infoValidation";
 import CustomButton from "../../components/CustomButton";
-import { replacePhone } from "../../utils/infoFormat";
+import { replacePhone } from "../../utils/format/infoFormat";
 import { getMyInfo, putAdminInfo } from "../../api/InfoAPI";
 
 type InfoAction = { type: "CHANGE"; payload: { key: string; value: string } } | { type: "RESET" };
