+++ src/utils/format/infoFormat.ts
